[Isolates] Merge 4831:4884 from bleeding_edge into isolates branch. Review URL: http://codereview.chromium.org/2871008
git-svn-id: http://v8.googlecode.com/svn/branches/experimental@4885 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ isolates/src/heap.h
@@ -1898,18 +1898,21 @@ class TranscendentalCache {
           return 0.0;  // Never happens.
       }
     }
+
     struct Element {
       uint32_t in[2];
       Object* output;
     };
+
     union Converter {
       double dbl;
       uint32_t integers[2];
     };
+
     inline static int Hash(const Converter& c) {
       uint32_t hash = (c.integers[0] ^ c.integers[1]);
-      hash ^= hash >> 16;
-      hash ^= hash >> 8;
+      hash ^= static_cast<int32_t>(hash) >> 16;
+      hash ^= static_cast<int32_t>(hash) >> 8;
       return (hash & (kCacheSize - 1));
     }
